+++ fltk-derive/src/display.rs
@@ -267,7 +267,7 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
             fn set_text_font(&mut self, font: Font) {
                 assert!(!self.was_deleted());
                 assert!(self.buffer().is_some());
-                unsafe { #set_text_font(self._inner, font as i32) }
+                unsafe { #set_text_font(self._inner, font.bits() as i32) }
             }
 
             fn text_color(&self) -> Color{
@@ -423,7 +423,7 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
                 let mut sizes: Vec<i32> = vec![];
                 for entry in entries.iter() {
                     colors.push(entry.color as u32);
-                    fonts.push(entry.font as i32);
+                    fonts.push(entry.font.bits() as i32);
                     sizes.push(entry.size as i32);
                 }
                 unsafe {
@@ -463,7 +463,7 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
             fn set_scrollbar_align(&mut self, align: Align){
                 unsafe {
                     assert!(!self.was_deleted());
-                    #set_scrollbar_align(self._inner, align as i32)
+                    #set_scrollbar_align(self._inner, align.bits() as i32)
                 }
             }
 
@@ -607,7 +607,7 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
             fn set_linenumber_font(&mut self, font: Font) {
                 unsafe {
                     assert!(!self.was_deleted());
-                    #set_linenumber_font(self._inner, font as i32)
+                    #set_linenumber_font(self._inner, font.bits() as i32)
                 }
             }
 
@@ -664,7 +664,7 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
             fn set_linenumber_align(&mut self, align: Align) {
                 unsafe {
                     assert!(!self.was_deleted());
-                    #set_linenumber_align(self._inner, align as i32)
+                    #set_linenumber_align(self._inner, align.bits() as i32)
                 }
             }