+++ fltk/src/draw.rs
@@ -495,7 +495,7 @@ pub fn end_complex_polygon() {
 
 /// Sets the current font, which is then used in various drawing routines
 pub fn set_font(face: Font, fsize: u32) {
-    unsafe { Fl_set_draw_font(face as i32, fsize as i32) }
+    unsafe { Fl_set_draw_font(face.bits() as i32, fsize as i32) }
 }
 
 /// Gets the current font, which is used in various drawing routines
@@ -516,7 +516,7 @@ pub fn height() -> i32 {
 /// Sets the line spacing for the current font
 pub fn set_height(font: Font, size: u32) {
     unsafe {
-        Fl_set_height(font as i32, size as i32);
+        Fl_set_height(font.bits() as i32, size as i32);
     }
 }
 
@@ -587,7 +587,7 @@ pub fn draw_text(txt: &str, x: i32, y: i32) {
 /// Draws a string starting at the given x, y location with width and height and alignment
 pub fn draw_text2(string: &str, x: i32, y: i32, width: i32, height: i32, align: Align) {
     let s = CString::safe_new(string);
-    unsafe { Fl_draw_text2(s.as_ptr(), x, y, width, height, align as i32) }
+    unsafe { Fl_draw_text2(s.as_ptr(), x, y, width, height, align.bits() as i32) }
 }
 
 /// Draws a string starting at the given x, y location, rotated to an angle
@@ -634,7 +634,7 @@ pub fn can_do_alpha_blending() -> bool {
 /// Get a human-readable string from a shortcut value
 pub fn shortcut_label(shortcut: Shortcut) -> String {
     unsafe {
-        let x = Fl_shortcut_label(shortcut as u32);
+        let x = Fl_shortcut_label(shortcut.bits() as u32);
         assert!(!x.is_null());
         CStr::from_ptr(x as *mut raw::c_char)
             .to_string_lossy()
@@ -672,7 +672,7 @@ pub fn set_spot<Win: WindowExt>(font: Font, size: u32, x: i32, y: i32, w: i32, h
     unsafe {
         assert!(!win.was_deleted());
         Fl_set_spot(
-            font as i32,
+            font.bits() as i32,
             size as i32,
             x,
             y,