+++ fltk/src/enums.rs
@@ -87,86 +87,54 @@ impl FrameType {
     }
 }
 
-/// Defines alignment rules used by FLTK for labels
-#[repr(i32)]
-#[derive(Copy, Clone, PartialEq)]
-pub enum Align {
-    Center = 0,
-    Top = 1,
-    Bottom = 2,
-    Left = 4,
-    Right = 8,
-    Inside = 16,
-    TextOverImage = 20,
-    Clip = 40,
-    Wrap = 80,
-    ImageNextToText = 100,
-    TextNextToImage = 120,
-    ImageBackdrop = 200,
-    TopLeft = 1 | 4,
-    TopRight = 1 | 8,
-    BottomLeft = 2 | 4,
-    BottomRight = 2 | 8,
-    LeftTop = 7,
-    RightTop = 11,
-    LeftBottom = 13,
-    RightBottom = 14,
-    PositionMask = 15,
-    ImageMask = 320,
-}
-
-#[allow(unreachable_patterns)]
-impl std::fmt::Debug for Align {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        use Align::*;
-        match *self {
-            Center => write!(f, "Align::Center"),
-            Top => write!(f, "Align::Top"),
-            Bottom => write!(f, "Align::Bottom"),
-            Left => write!(f, "Align::Left"),
-            Right => write!(f, "Align::Right"),
-            Inside => write!(f, "Align::Inside"),
-            TextOverImage => write!(f, "Align::TextOverImage"),
-            Clip => write!(f, "Align::Clip"),
-            Wrap => write!(f, "Align::Wrap"),
-            ImageNextToText => write!(f, "Align::ImageNextToText"),
-            TextNextToImage => write!(f, "Align::TextNextToImage"),
-            ImageBackdrop => write!(f, "Align::ImageBackdrop"),
-            TopLeft => write!(f, "Align::TopLeft"),
-            TopRight => write!(f, "Align::TopRight"),
-            BottomLeft => write!(f, "Align::BottomLeft"),
-            BottomRight => write!(f, "Align::BottomRight"),
-            LeftTop => write!(f, "Align::LeftTop"),
-            RightTop => write!(f, "Align::RightTop"),
-            LeftBottom => write!(f, "Align::LeftBottom"),
-            RightBottom => write!(f, "Align::RightBottom"),
-            PositionMask => write!(f, "Align::PositionMask"),
-            ImageMask => write!(f, "Align::ImageMask"),
-            _ => write!(f, "0x{:02x}", *self as i32),
-        }
+bitflags! {
+    /// Defines alignment rules used by FLTK for labels
+    pub struct Align: i32 {
+        const Center = 0;
+        const Top = 1;
+        const Bottom = 2;
+        const Left = 4;
+        const Right = 8;
+        const Inside = 16;
+        const TextOverImage = 20;
+        const Clip = 40;
+        const Wrap = 80;
+        const ImageNextToText = 100;
+        const TextNextToImage = 120;
+        const ImageBackdrop = 200;
+        const TopLeft = 1 | 4;
+        const TopRight = 1 | 8;
+        const BottomLeft = 2 | 4;
+        const BottomRight = 2 | 8;
+        const LeftTop = 7;
+        const RightTop = 11;
+        const LeftBottom = 13;
+        const RightBottom = 14;
+        const PositionMask = 15;
+        const ImageMask = 320;
     }
 }
 
-/// Defines fonts used by FLTK
-#[repr(i32)]
-#[derive(Copy, Clone, PartialEq)]
-pub enum Font {
-    Helvetica = 0,
-    HelveticaBold = 1,
-    HelveticaItalic = 2,
-    HelveticaBoldItalic = 3,
-    Courier = 4,
-    CourierBold = 5,
-    CourierItalic = 6,
-    CourierBoldItalic = 7,
-    Times = 8,
-    TimesBold = 9,
-    TimesItalic = 10,
-    TimesBoldItalic = 11,
-    Symbol = 12,
-    Screen = 13,
-    ScreenBold = 14,
-    Zapfdingbats = 15,
+bitflags! {
+    /// Defines fonts used by FLTK
+    pub struct Font: i32 {
+        const Helvetica = 0;
+        const HelveticaBold = 1;
+        const HelveticaItalic = 2;
+        const HelveticaBoldItalic = 3;
+        const Courier = 4;
+        const CourierBold = 5;
+        const CourierItalic = 6;
+        const CourierBoldItalic = 7;
+        const Times = 8;
+        const TimesBold = 9;
+        const TimesItalic = 10;
+        const TimesBoldItalic = 11;
+        const Symbol = 12;
+        const Screen = 13;
+        const ScreenBold = 14;
+        const Zapfdingbats = 15;
+    }
 }
 
 impl Font {
@@ -186,33 +154,6 @@ impl Font {
         match font_index(name) {
             Some(val) => Font::by_index(val),
             None => Font::Helvetica,
-        }
-    }
-}
-
-
-#[allow(unreachable_patterns)]
-impl std::fmt::Debug for Font {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        use Font::*;
-        match *self {
-            Helvetica => write!(f, "Font::Helvetica"),
-            HelveticaBold => write!(f, "Font::HelveticaBold"),
-            HelveticaItalic => write!(f, "Font::HelveticaItalic"),
-            HelveticaBoldItalic => write!(f, "Font::HelveticaBoldItalic"),
-            Courier => write!(f, "Font::Courier"),
-            CourierBold => write!(f, "Font::CourierBold"),
-            CourierItalic => write!(f, "Font::CourierItalic"),
-            CourierBoldItalic => write!(f, "Font::CourierBoldItalic"),
-            Times => write!(f, "Font::Times"),
-            TimesBold => write!(f, "Font::TimesBold"),
-            TimesItalic => write!(f, "Font::TimesItalic"),
-            TimesBoldItalic => write!(f, "Font::TimesBoldItalic"),
-            Symbol => write!(f, "Font::Symbol"),
-            Screen => write!(f, "Font::Screen"),
-            ScreenBold => write!(f, "Font::ScreenBold"),
-            Zapfdingbats => write!(f, "Font::Zapfdingbats"),
-            _ => write!(f, "0x{:02x}", *self as i32),
         }
     }
 }
@@ -339,51 +280,53 @@ pub enum Event {
     ZoomEvent,
 }
 
-/// Defines the inputted virtual keycode
-#[repr(i32)]
-#[derive(Copy, Clone, PartialEq)]
-pub enum Key {
-    None = 0,
-    Button = 0xfee8,
-    BackSpace = 0xff08,
-    Tab = 0xff09,
-    IsoKey = 0xff0c,
-    Enter = 0xff0d,
-    Pause = 0xff13,
-    ScrollLock = 0xff14,
-    Escape = 0xff1b,
-    Kana = 0xff2e,
-    Eisu = 0xff2f,
-    Yen = 0xff30,
-    JISUnderscore = 0xff31,
-    Home = 0xff50,
-    Left = 0xff51,
-    Up = 0xff52,
-    Right = 0xff53,
-    Down = 0xff54,
-    PageUp = 0xff55,
-    PageDown = 0xff56,
-    End = 0xff57,
-    Print = 0xff61,
-    Insert = 0xff63,
-    Menu = 0xff67,
-    Help = 0xff68,
-    NumLock = 0xff7f,
-    KP = 0xff80,
-    KPEnter = 0xff8d,
-    KPLast = 0xffbd,
-    FLast = 0xffe0,
-    ShiftL = 0xffe1,
-    ShiftR = 0xffe2,
-    ControlL = 0xffe3,
-    ControlR = 0xffe4,
-    CapsLock = 0xffe5,
-    MetaL = 0xffe7,
-    MetaR = 0xffe8,
-    AltL = 0xffe9,
-    AltR = 0xffea,
-    Delete = 0xffff,
+bitflags! {
+    /// Defines the inputted virtual keycode
+    pub struct Key: i32 {
+        const None = 0;
+        const Button = 0xfee8;
+        const BackSpace = 0xff08;
+        const Tab = 0xff09;
+        const IsoKey = 0xff0c;
+        const Enter = 0xff0d;
+        const Pause = 0xff13;
+        const ScrollLock = 0xff14;
+        const Escape = 0xff1b;
+        const Kana = 0xff2e;
+        const Eisu = 0xff2f;
+        const Yen = 0xff30;
+        const JISUnderscore = 0xff31;
+        const Home = 0xff50;
+        const Left = 0xff51;
+        const Up = 0xff52;
+        const Right = 0xff53;
+        const Down = 0xff54;
+        const PageUp = 0xff55;
+        const PageDown = 0xff56;
+        const End = 0xff57;
+        const Print = 0xff61;
+        const Insert = 0xff63;
+        const Menu = 0xff67;
+        const Help = 0xff68;
+        const NumLock = 0xff7f;
+        const KP = 0xff80;
+        const KPEnter = 0xff8d;
+        const KPLast = 0xffbd;
+        const FLast = 0xffe0;
+        const ShiftL = 0xffe1;
+        const ShiftR = 0xffe2;
+        const ControlL = 0xffe3;
+        const ControlR = 0xffe4;
+        const CapsLock = 0xffe5;
+        const MetaL = 0xffe7;
+        const MetaR = 0xffe8;
+        const AltL = 0xffe9;
+        const AltR = 0xffea;
+        const Delete = 0xffff;
+    }
 }
+
+
 
 impl Key {
     /// Gets a Key from an i32
@@ -397,65 +340,17 @@ impl Key {
     }
 }
 
-#[allow(unreachable_patterns)]
-impl std::fmt::Debug for Key {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match *self {
-            Key::None => write!(f, "Key::None"),
-            Key::Button => write!(f, "Key::Button"),
-            Key::BackSpace => write!(f, "Key::BackSpace"),
-            Key::Tab => write!(f, "Key::Tab"),
-            Key::IsoKey => write!(f, "Key::IsoKey"),
-            Key::Enter => write!(f, "Key::Enter"),
-            Key::Pause => write!(f, "Key::Pause"),
-            Key::ScrollLock => write!(f, "Key::ScrollLock"),
-            Key::Escape => write!(f, "Key::Escape"),
-            Key::Kana => write!(f, "Key::Kana"),
-            Key::Eisu => write!(f, "Key::Eisu"),
-            Key::Yen => write!(f, "Key::Yen"),
-            Key::JISUnderscore => write!(f, "Key::JISUnderscore"),
-            Key::Home => write!(f, "Key::Home"),
-            Key::Left => write!(f, "Key::Left"),
-            Key::Up => write!(f, "Key::Up"),
-            Key::Right => write!(f, "Key::Right"),
-            Key::Down => write!(f, "Key::Down"),
-            Key::PageUp => write!(f, "Key::PageUp"),
-            Key::PageDown => write!(f, "Key::PageDown"),
-            Key::End => write!(f, "Key::End"),
-            Key::Print => write!(f, "Key::Print"),
-            Key::Insert => write!(f, "Key::Insert"),
-            Key::Menu => write!(f, "Key::Menu"),
-            Key::Help => write!(f, "Key::Help"),
-            Key::NumLock => write!(f, "Key::NumLock"),
-            Key::KP => write!(f, "Key::KP"),
-            Key::KPEnter => write!(f, "Key::KPEnter"),
-            Key::KPLast => write!(f, "Key::KPLast"),
-            Key::FLast => write!(f, "Key::FLast"),
-            Key::ShiftL => write!(f, "Key::ShiftL"),
-            Key::ShiftR => write!(f, "Key::ShiftR"),
-            Key::ControlL => write!(f, "Key::ControlL"),
-            Key::ControlR => write!(f, "Key::ControlR"),
-            Key::CapsLock => write!(f, "Key::CapsLock"),
-            Key::MetaL => write!(f, "Key::MetaL"),
-            Key::MetaR => write!(f, "Key::MetaR"),
-            Key::AltL => write!(f, "Key::AltL"),
-            Key::AltR => write!(f, "Key::AltR"),
-            Key::Delete => write!(f, "Key::Delete"),
-            _ => write!(f, "0x{:02x}", *self as i32),
-        }
+bitflags! {
+    /// Defines the modifiers of virtual keycodes
+    pub struct Shortcut: i32 {
+        const None = 0;
+        const Shift = 0x0001_0000;
+        const CapsLock = 0x0002_0000;
+        const Ctrl = 0x0004_0000;
+        const Alt = 0x0008_0000;
     }
 }
 
-/// Defines the modifiers of virtual keycodes
-#[repr(i32)]
-#[derive(Copy, Clone, PartialEq)]
-pub enum Shortcut {
-    None = 0,
-    Shift = 0x0001_0000,
-    CapsLock = 0x0002_0000,
-    Ctrl = 0x0004_0000,
-    Alt = 0x0008_0000,
-}
 
 impl Shortcut {
     /// Create a shortcut from a char
@@ -474,53 +369,20 @@ impl Shortcut {
     }
 }
 
-#[allow(unreachable_patterns)]
-impl std::fmt::Debug for Shortcut {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        use Shortcut::*;
-        match *self {
-            None => write!(f, "Shortcut::None"),
-            Shift => write!(f, "Shortcut::Shift"),
-            CapsLock => write!(f, "Shortcut::CapsLock"),
-            Ctrl => write!(f, "Shortcut::Ctrl"),
-            Alt => write!(f, "Shortcut::Alt"),
-            _ => write!(f, "0x{:08x}", *self as i32),
-        }
+bitflags! {
+    /// Defines the types of triggers for widget callback functions
+    pub struct CallbackTrigger: i32 {
+        const Never = 0;
+        const Changed = 1;
+        const NotChanged = 2;
+        const Release = 4;
+        const ReleaseAlways = 6;
+        const EnterKey = 8;
+        const EnterKeyAlways = 10;
+        const EnterKeyChanged = 11;
     }
 }
 
-/// Defines the types of triggers for widget callback functions
-#[repr(i32)]
-#[derive(Copy, Clone, PartialEq)]
-pub enum CallbackTrigger {
-    Never = 0,
-    Changed = 1,
-    NotChanged = 2,
-    Release = 4,
-    ReleaseAlways = 6,
-    EnterKey = 8,
-    EnterKeyAlways = 10,
-    EnterKeyChanged = 11,
-}
-
-
-#[allow(unreachable_patterns)]
-impl std::fmt::Debug for CallbackTrigger {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        use CallbackTrigger::*;
-        match *self {
-            Never => write!(f, "CallbackTrigger::Never"),
-            Changed => write!(f, "CallbackTrigger::Changed"),
-            NotChanged => write!(f, "CallbackTrigger::NotChanged"),
-            Release => write!(f, "CallbackTrigger::Release"),
-            ReleaseAlways => write!(f, "CallbackTrigger::ReleaseAlways"),
-            EnterKey => write!(f, "CallbackTrigger::EnterKey"),
-            EnterKeyAlways => write!(f, "CallbackTrigger::EnterKeyAlways"),
-            EnterKeyChanged => write!(f, "CallbackTrigger::EnterKeyChanged"),
-            _ => write!(f, "0x{:02x}", *self as i32),
-        }
-    }
-}
 
 /// Defines the text cursor styles supported by fltk
 #[repr(i32)]
@@ -561,44 +423,21 @@ pub enum Cursor {
     None = 255,
 }
 
-/// Defines Fl_Mode types
-#[repr(i32)]
-#[derive(Copy, Clone, PartialEq)]
-pub enum Mode {
-    Rgb = 0,
-    Index = 1,
-    Double = 2,
-    Accum = 4,
-    Alpha = 8,
-    Depth = 16,
-    Stencil = 32,
-    Rgb8 = 64,
-    MultiSample = 128,
-    Stereo = 256,
-    FakeSingle = 512, // Fake single buffered windows using double-buffer
-    Opengl3 = 1024,
-}
-
-
-#[allow(unreachable_patterns)]
-impl std::fmt::Debug for Mode {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        use Mode::*;
-        match *self {
-            Rgb => write!(f, "Mode::Rgb"),
-            Index => write!(f, "Mode::Index"),
-            Double => write!(f, "Mode::Double"),
-            Accum => write!(f, "Mode::Accum"),
-            Alpha => write!(f, "Mode::Alpha"),
-            Depth => write!(f, "Mode::Depth"),
-            Stencil => write!(f, "Mode::Stencil"),
-            Rgb8 => write!(f, "Mode::Rgb8"),
-            MultiSample => write!(f, "Mode::MultiSample"),
-            Stereo => write!(f, "Mode::Stereo"),
-            FakeSingle => write!(f, "Mode::FakeSingle"),
-            Opengl3 => write!(f, "Mode::Opengl3"),
-            _ => write!(f, "0x{:04x}", *self as i32),
-        }
+bitflags! {
+    /// Defines Fl_Mode types
+    pub struct Mode: i32 {
+        const Rgb = 0;
+        const Index = 1;
+        const Double = 2;
+        const Accum = 4;
+        const Alpha = 8;
+        const Depth = 16;
+        const Stencil = 32;
+        const Rgb8 = 64;
+        const MultiSample = 128;
+        const Stereo = 256;
+        const FakeSingle = 512; // Fake single buffered windows using double-buffer
+        const Opengl3 = 1024;
     }
 }
 
@@ -610,69 +449,20 @@ pub trait WidgetType {
 impl std::ops::BitOr<char> for Shortcut {
     type Output = Shortcut;
     fn bitor(self, other: char) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | other as i32) }
+        unsafe { std::mem::transmute(self.bits as i32 | other as i32) }
     }
 }
 
 impl std::ops::BitOr<Key> for Shortcut {
     type Output = Shortcut;
     fn bitor(self, other: Key) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | other as i32) }
-    }
-}
-
-impl std::ops::BitOr<Shortcut> for Shortcut {
-    type Output = Shortcut;
-    fn bitor(self, other: Shortcut) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | other as i32) }
-    }
-}
-
-impl std::ops::BitOr<CallbackTrigger> for CallbackTrigger {
-    type Output = CallbackTrigger;
-    fn bitor(self, rhs: CallbackTrigger) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | rhs as i32) }
-    }
-}
-
-impl std::ops::BitOr<Align> for Align {
-    type Output = Align;
-    fn bitor(self, rhs: Align) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | rhs as i32) }
+        unsafe { std::mem::transmute(self.bits as i32 | other.bits() as i32) }
     }
 }
 
 impl std::ops::BitOr<i32> for Align {
     type Output = Align;
     fn bitor(self, rhs: i32) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | rhs as i32) }
-    }
-}
-
-impl std::ops::BitOr<Font> for Font {
-    type Output = Font;
-    fn bitor(self, rhs: Font) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | rhs as i32) }
-    }
-}
-
-impl std::ops::BitOr<Event> for Event {
-    type Output = Event;
-    fn bitor(self, rhs: Event) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | rhs as i32) }
-    }
-}
-
-impl std::ops::BitOr<Key> for Key {
-    type Output = Key;
-    fn bitor(self, rhs: Key) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | rhs as i32) }
-    }
-}
-
-impl std::ops::BitOr<Mode> for Mode {
-    type Output = Mode;
-    fn bitor(self, rhs: Mode) -> Self::Output {
-        unsafe { std::mem::transmute(self as i32 | rhs as i32) }
+        unsafe { std::mem::transmute(self.bits | rhs as i32) }
     }
 }