+++ fltk/src/tree.rs
@@ -311,7 +311,7 @@ impl Tree {
             TreeItem::from_raw(Fl_Tree_next_visible_item(
                 self._inner,
                 start._inner,
-                direction_key as i32,
+                direction_key.bits() as i32,
             ))
         }
     }
@@ -343,7 +343,7 @@ impl Tree {
             TreeItem::from_raw(Fl_Tree_next_item(
                 self._inner,
                 item._inner,
-                direction_key as i32,
+                direction_key.bits() as i32,
                 visible as i32,
             ))
         }
@@ -359,7 +359,7 @@ impl Tree {
             TreeItem::from_raw(Fl_Tree_next_selected_item(
                 self._inner,
                 item._inner,
-                direction_key as i32,
+                direction_key.bits() as i32,
             ))
         }
     }
@@ -562,7 +562,7 @@ impl Tree {
                 self._inner,
                 from._inner,
                 to._inner,
-                direction_key as i32,
+                direction_key.bits() as i32,
                 val as i32,
                 visible as i32,
             ) {
@@ -632,7 +632,7 @@ impl Tree {
     /// Sets the items' label font
     pub fn set_item_label_font(&mut self, val: Font) {
         assert!(!self.was_deleted());
-        unsafe { Fl_Tree_set_item_labelfont(self._inner, val as i32) }
+        unsafe { Fl_Tree_set_item_labelfont(self._inner, val.bits() as i32) }
     }
 
     /// Gets the items' label size
@@ -1297,7 +1297,7 @@ impl TreeItem {
     /// Sets the label's font
     pub fn set_label_font(&mut self, val: Font) {
         assert!(!self.was_deleted());
-        unsafe { Fl_Tree_Item_set_labelfont(self._inner, val as i32) }
+        unsafe { Fl_Tree_Item_set_labelfont(self._inner, val.bits() as i32) }
     }
 
     /// Gets the label's font