+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.11.4] - Unreleased
 ### Changes
 - Use lazy_static for globals.
+- Use bitflags for several enum types.
 
 ## [0.11.3] - 2020-11-30
 ### Changes